+++ tests/test_signals.py
@@ -1,15 +1,4 @@
-import pytest
-
-try:
-    import blinker
-except ImportError:
-    blinker = None
-
 import flask
-
-pytestmark = pytest.mark.skipif(
-    blinker is None, reason="Signals require the blinker library."
-)
 
 
 def test_template_rendered(app, client):